pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "indirectrates"
version = "0.1.0"
description = "Indirect rate forecasting agent (GovCon-oriented MVP)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.1",
  "numpy>=1.26",
  "PyYAML>=6.0",
  "matplotlib>=3.8",
  "openpyxl>=3.1",
  "typer>=0.12",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]
server = ["fastapi>=0.115", "uvicorn[standard]>=0.30", "python-multipart>=0.0.9"]

[project.scripts]
indirectrates = "indirectrates.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
indirectrates = ["resources/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
